RavenDB-22955 Replication - Usability: Add Replication Progress
