+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/partials/OngoingEtlTaskProgressTooltip.tsx
