+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/partials/OngoingTaskProgressProvider.tsx
