+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/partials/OngoingTaskSelectActions.tsx
@@ -13,7 +13,7 @@ import {
     DropdownItem,
     Button,
 } from "reactstrap";
-import { OngoingTaskOperationConfirmType } from "../shared/OngoingTaskOperationConfirm";
+import { OngoingTaskOperationConfirmType } from "../../shared/OngoingTaskOperationConfirm";
 
 interface OngoingTaskSelectActionsProps {
     allTasks: number[];
